- Summary: fix bug140623,137874 - Bugzilla Bug (s) Resolved: 140623,137874 - Description: 1. use automatic algorithm to detect 3d chart size. 2. remove spacing between 3d axes and floor, remove tick rendering for 3d axes. 3. remove usless import entry and type casting in some files - Tests Description  : Manual test - Notes to Build Team:
- Notes to Developers:

- Notes to QA:

- Notes to Documentation:

- Files Edited:
@changelist
- Files Added:
@addlist
- Files Deleted
@deletelist

+++ chart/org.eclipse.birt.chart.engine/src/org/eclipse/birt/chart/computation/withaxes/PlotWith2DAxes.java
@@ -1803,7 +1803,7 @@ public final class PlotWith2DAxes extends PlotWithAxes
 				}
 				else if ( isZeroValue == true )
 				{
-					percentileValue = new Double ( (double)1.0 / iOrthogonalCount );
+					percentileValue = new Double ( 1.0 / iOrthogonalCount );
 				}
 
 				dpa[i] = new DataPointHints( oDataBase,
